backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
yxdb==1.1.1
google-generativeai==0.3.2
aiofiles==23.2.1
python-dotenv==1.0.0
